+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All Notable changes to `bytenet\laravel-admn-base` project will be documented in
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2016-12-28
+### Added
+- src/day02/README.md
+- src/day02/NumberNames.php
+- tests/day02/NumberNamesTest.php
+
+### Changed
+- .travis.yml
+
 ## [0.1.1] - 2016-12-28
 ### Changed
 - .travis.yml file
@@ -21,4 +30,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 -  Day-01 - SRC and TEST
 
+[0.2.0]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.1.0...v0.1.1